Reworking event dispatch registration system.
Now when subscribing event listeners, a unique key idenifying their registration is created and returned. This key can be used later on to remove the registration of a listener from the event dispatcher.

This is useful to remove event listeners of non-equatable types, and also for the new ClosureEventListener helper struct that was created.